lab5/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beanie==1.29.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.11
h11==0.14.0
idna==3.10
lazy-model==0.2.0
motor==3.7.0
pydantic==2.10.6
pydantic-mongo==3.1.0
pydantic_core==2.27.2
pymongo==4.12.0
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.1
toml==0.10.2
typing_extensions==4.12.2
uvicorn==0.34.0
